// frontend/gear-vault-web/src/api/generated/client/models/CreatePartFailurePredictionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreatePartFailurePredictionDto = {
    currentMileage?: number | null;
    averageDailyKilometers?: number | null;
    lastServiceDate?: string | null;
    conditionNotes?: string | null;
    usagePattern?: string | null;
};